chore: ignore observer errors in cypress
