+++ e2e/cypress/support/e2e.ts
@@ -35,9 +35,9 @@ Cypress.on('uncaught:exception', (err, runnable) => {
   if (err.hasOwnProperty('code') && typeof err.code == 'string') {
     return false;
   }
-  // resize observer throwing random errors, hard to replicate, doesn't seem to affect anything
-  const resizeObserverLoopErrRe = /^ResizeObserver loop limit exceeded/;
-  if (resizeObserverLoopErrRe.test(err.message)) {
+
+  // resize observer throwing these errors sometimes, doesn't seem to affect anything
+  if (err.message.includes('ResizeObserver loop')) {
     return false;
   }
 });
