+++ webapp/src/globalContext/useLayoutService.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import type { useQuickStartGuideService } from './useQuickStartGuideService';
-import { useResizeObserver } from 'tg.hooks/useResizeObserver';
+import { useWidthObserver } from 'tg.hooks/useWidthObserver';
 
 export const TOP_BAR_HEIGHT = 52;
 
@@ -14,12 +14,12 @@ export const useLayoutService = ({ quickStart }: Props) => {
   const [topBannerHeight, setTopBannerHeight] = useState(0);
   const [topSubBannerHeight, setTopSubBannerHeight] = useState(0);
   const [topBarHidden, setTopBarHidden] = useState(false);
-  const bodySize = useResizeObserver({
-    ref: {
-      current: typeof document !== 'undefined' ? document.body : null,
-    },
-  });
-  const bodyWidth = bodySize.width ?? 0;
+  const bodyWidth =
+    useWidthObserver({
+      ref: {
+        current: typeof document !== 'undefined' ? document.body : null,
+      },
+    }) ?? 0;
 
   const [rightPanelFloatingForced, setRightPanelFloatingForced] =
     useState(false);
